Fix/#175 fix member delete error (#178)
* fix: Base64관련 디코딩 코드 변경 -> Base64Url

* refactor: 쿠폰 스케쥴 업데이트 및 config 수정

* style: 문자열 checkstyle 수정

* fix: 회원 탈퇴시 방 참여에 대한 문제 해결

* refactor: config update

* test: 신고 실패에 대한 테스트 코드 변경

+++ src/main/java/com/moabam/api/application/member/MemberService.java
@@ -17,6 +17,9 @@ import com.moabam.api.domain.item.repository.ItemRepository;
 import com.moabam.api.domain.member.Member;
 import com.moabam.api.domain.member.repository.MemberRepository;
 import com.moabam.api.domain.member.repository.MemberSearchRepository;
+import com.moabam.api.domain.room.Participant;
+import com.moabam.api.domain.room.repository.ParticipantRepository;
+import com.moabam.api.domain.room.repository.ParticipantSearchRepository;
 import com.moabam.api.dto.auth.AuthorizationTokenInfoResponse;
 import com.moabam.api.dto.auth.LoginResponse;
 import com.moabam.api.dto.member.MemberInfo;
@@ -42,6 +45,8 @@ public class MemberService {
 	private final InventoryRepository inventoryRepository;
 	private final ItemRepository itemRepository;
 	private final MemberSearchRepository memberSearchRepository;
+	private final ParticipantSearchRepository participantSearchRepository;
+	private final ParticipantRepository participantRepository;
 	private final ClockHolder clockHolder;
 
 	public Member findMember(Long memberId) {
@@ -69,6 +74,12 @@ public class MemberService {
 
 	@Transactional
 	public void delete(Member member) {
+		List<Participant> participants = participantRepository.findAllByMemberId(member.getId());
+
+		if (!participants.isEmpty()) {
+			throw new BadRequestException(NEED_TO_EXIT_ALL_ROOMS);
+		}
+
 		member.delete(clockHolder.times());
 		memberRepository.flush();
 		memberRepository.delete(member);
@@ -92,14 +103,29 @@ public class MemberService {
 		Member member = memberSearchRepository.findMember(authMember.id())
 			.orElseThrow(() -> new NotFoundException(MEMBER_NOT_FOUND));
 
-		member.changeNickName(modifyMemberRequest.nickname());
+		boolean nickNameChanged = member.changeNickName(modifyMemberRequest.nickname());
 		member.changeIntro(modifyMemberRequest.intro());
 		member.changeProfileUri(newProfileUri);
 
 		memberRepository.save(member);
+
+		if (nickNameChanged) {
+			changeNickname(authMember.id(), modifyMemberRequest.nickname());
+		}
+	}
+
+	private void changeNickname(Long memberId, String changedName) {
+		List<Participant> participants = participantSearchRepository.findAllRoomMangerByMemberId(memberId);
+
+		for (Participant participant : participants) {
+			participant.getRoom().changeManagerNickname(changedName);
+		}
 	}
 
 	private void validateNickname(String nickname) {
+		if (Objects.isNull(nickname)) {
+			return;
+		}
 		if (StringUtils.isEmpty(nickname) && memberRepository.existsByNickname(nickname)) {
 			throw new ConflictException(NICKNAME_CONFLICT);
 		}
